Update user agent version automatically if it is outdated.
On regular intervalls WhatsApp updates its client version and requires it on registrations. If the sent version is outdated, a registration is not possible. Until now, the user agent of CoSeMe had to be updated manually by the developers. This is pretty annoying for developers and for users.

So this commit integrates a mechanism which tries to get the latest WhatsApp version from a third party service if it is needed and uses it when the config is built.